// kservices_ts_api/src/api/domain/entities/v3/sme.entity.ts
export type fetchRequest = {
  id: number;
};

export type tokenRequest = {
  id: string;
  token: string;
};

export type fetchLeadCustomerDetailRequest = {
  id: number;
  customerNumber: string;
  offset: number;
  mode: string;
};
// ...
export type updateLeadSettingsRequest = {
  id : number;
  leadColumns:string;
};
// ...
export type getLeadSourceRequest = {
  smeId: number;
};

export type getLeadStatusRequest = {
  smeId: number;
};